feat(dashmate): prebuild services dependencies
